% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brightnessmask.R
\name{brightness.mask}
\alias{brightness.mask}
\title{Convert a hdf5 file containing hyperspectral imagery into a brightness mask for use in later corrections.}
\usage{
brightness.mask(hy.file, metadata.path, reflectance.path, wavelength.path)
}
\arguments{
\item{hy.file}{hdf5 file containing hyperspectral imagery and associated metadata}

\item{metadata.path}{hdf5 path to reflectance metadata}

\item{reflectance.path}{hdf5 path to reflectance data}

\item{wavelength.path}{hdf5 path to wavelength metadata}
}
\value{
A matrix containing the brightness mask
}
\description{
This function reads in a hdf5 file and create a brightness matrix to remove shaded pixels for use in the 
BRDF correction that is applied in later steps.
}
\details{
For theory behind this function please see:

Clark, M.L., Roberts, D.A., and Clark, D.B., 2005. Hyperspectral discrimination of tropical rain forest
tree species at crown scales. Remote Sensing of Environment, 96: 375-398.

Gougeon, F.A., 1995. Comparison of possible multispectral classification schemes for tree crowns individually
delineated on high spatial resolution MEIS images. Canadian Journal of Remote Sensing, 21(1): 1-9.
}
