% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsicoef.R
\name{hsi.coef}
\alias{hsi.coef}
\title{This function first reads in a corrected hdf5 file and applies a supplied coefficient to each band.
Coefficients may be from a PLSR, PCA, etc. Output is a raster file of same spatial resolution as input.}
\usage{
hsi.coef(hy.file, metadata.path, reflectance.path, wavelength.path,
  coordinate.path, coef.csv, inter, scale.data)
}
\arguments{
\item{hy.file}{hdf5 file containing hyperspectral imagery and associated metadata}

\item{metadata.path}{hdf5 path to reflectance metadata}

\item{reflectance.path}{hdf5 path to reflectance data}

\item{wavelength.path}{hdf5 path to wavelength metadata}

\item{coordinate.path}{hdf5 path to coordinate data}

\item{coef.csv}{csv containing coefficients}

\item{inter}{does the CSV contain an intercept and is it in the first position (TRUE/FALSE)}

\item{scale.data}{were the coefficients generated with scaled data and thus data needs to be scaled before applied coefs (TRUE/FALSE)}
}
\value{
A raster with all coefficients applied
}
\description{
This function first reads in a corrected hdf5 file and applies a supplied coefficient to each band.
Coefficients may be from a PLSR, PCA, etc. Output is a raster file of same spatial resolution as input.
}
