% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsirandomextract.R
\name{hsi.random.extract}
\alias{hsi.random.extract}
\title{Extract refelctance data from a set of random points from a topographically and brdf corrected hdf5 file
containing hyperspectral imagery.}
\usage{
hsi.random.extract(hy.file, coordinate.path, wavelength.path, band.combo,
  number.pts)
}
\arguments{
\item{hy.file}{hdf5 file containing corrected hyperspectral imagery}

\item{coordinate.path}{hdf5 path to the coordinate information}

\item{wavelength.path}{hdf5 path to the wavelength information}

\item{band.combo}{combination of wavelengths you want to extract data from}

\item{number.pts}{number of random points you want to extract from the file}
}
\value{
A matrix topographic and brdf corrected reflectance data
}
\description{
Extract refelctance data from a set of random points from a topographically and brdf corrected hdf5 file
containing hyperspectral imagery.
}
