% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsishpextract.R
\name{hsi.shp.extract}
\alias{hsi.shp.extract}
\title{Extract refelctance data from a set of random points from a topographically and brdf corrected hdf5 file
containing hyperspectral imagery.}
\usage{
hsi.shp.extract(hy.file, metadata.path, reflectance.path, coordinate.path,
  wavelength.path, band.combo, shp.file.loc, shp.file.name)
}
\arguments{
\item{hy.file}{hdf5 file containing corrected hyperspectral imagery}

\item{metadata.path}{hdf5 path to the metadata}

\item{reflectance.path}{hdf5 path to the reflectance data}

\item{coordinate.path}{hdf5 path to the coordinate information}

\item{wavelength.path}{hdf5 path to the wavelength information}

\item{band.combo}{combination of wavelengths you want to extract data from}

\item{shp.file.loc}{file path to shape file containing point data to extract refelctance data}

\item{shp.file.name}{name of shape file}
}
\value{
A matrix topographic and brdf corrected reflectance data
}
\description{
Extract refelctance data from a set of random points from a topographically and brdf corrected hdf5 file
containing hyperspectral imagery.
}
