% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvimask.R
\name{ndvi.mask}
\alias{ndvi.mask}
\title{Convert a hdf5 file containing hyperspectral imagery into a NDVI mask for use in later corrections.}
\usage{
ndvi.mask(hy.file, metadata.path, reflectance.path, wavelength.path,
  red.nm, nir.nm, ndvi.threshold)
}
\arguments{
\item{hy.file}{hdf5 file containing hyperspectral imagery and associated metadata}

\item{metadata.path}{hdf5 path to reflectance metadata}

\item{reflectance.path}{hdf5 path to reflectance data}

\item{wavelength.path}{hdf5 path to wavelength metadata}

\item{red.nm}{wavelength (nm) to use for red band}

\item{nir.nm}{wavelength (nm) to use for nir band}

\item{ndvi.threshold}{threshold for NDVI mask}
}
\value{
A matrix containing the NDVI mask
}
\description{
This function reads in a hdf5 file and create a NDVI matrix for use in the topographic and BRDF corrections that are
applied in later steps.
}
