% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topocorrection.R
\name{topo.correction}
\alias{topo.correction}
\title{Convert a hdf5 file containing hyperspectral imagery into a topographically corrected reflectance matrix.}
\usage{
topo.correction(hy.file, ndvi.mask, metadata.path, reflectance.path,
  wavelength.path, solar.az.path, solar.zn.path, slope.path, aspect.path,
  band.combo)
}
\arguments{
\item{hy.file}{hdf5 file containing hyperspectral imagery and associated metadata}

\item{ndvi.mask}{NDVI mask created with the ndvi.mask function}

\item{metadata.path}{hdf5 path to reflectance metadata}

\item{reflectance.path}{hdf5 path to reflectance data}

\item{wavelength.path}{hdf5 path to wavelength metadata}

\item{solar.az.path}{hdf5 path to solar azimuth data}

\item{solar.zn.path}{hdf5 path to solar zenith data}

\item{slope.path}{hdf5 path to slope data}

\item{aspect.path}{hdf5 path to aspect data}

\item{band.combo}{index of bands that will be processed}
}
\value{
A list of topographically corrected matrices
}
\description{
This function reads in a hdf5 file and create a topographically corrected reflectance matrix.
}
\details{
For theory behind this function please see:

Soenen, S.A., Peddle, D.R., and Coburn, C.A., 2005. SCS+C: A Modified Sun-Canopy-Sensor Topographic Correction 
in Forested Terrain. IEEE Transactions on Geoscience and Remote Sensing, 43(9): 2148-2159.
}
